Completed: Library display, Add book form
This is a late push. I have completed the base html/css structure for
the webpage. The library display shows all the current books in user'
library.

They also have the ability to add a new book using the
"Add Book" button located at the header of the webpage. This allows
them to add the book and it will create a new book in the myLibrary
array as well as add it to the library display.

Currently working on removing a book from the library.